Refactor fylke-validering
