+++ src/main/resources/lib/search/external/update-all.ts
@@ -4,7 +4,10 @@ import { logger } from '../../utils/logging';
 import { forceArray } from '../../utils/array-utils';
 import { getRepoConnection } from '../../utils/repo-utils';
 import { queryAllLayersToLocaleBuckets } from '../../localization/locale-utils';
-import { buildExternalSearchDocument, ExternalSearchDocument } from './document-builder';
+import {
+    buildExternalSearchDocument,
+    ExternalSearchDocument,
+} from './document-builder/document-builder';
 import { getLayersData } from '../../localization/layers-data';
 import { searchApiPostDocumentsAsync } from './api-handlers/post-document';
